extract common function to util package
