+++ client/starwhale/api/_impl/metric.py
@@ -13,6 +13,8 @@ from sklearn.metrics import (  # type: ignore
     multilabel_confusion_matrix,
 )
 
+from starwhale.utils.flatten import do_flatten_dict
+
 
 class MetricKind:
     MultiClassification = "multi_classification"
@@ -65,44 +67,11 @@ def multi_classification(
                     _r["roc_auc"][_label] = _calculate_roc_auc(
                         y_true, y_pr, _label, _idx
                     )
-            return _do_flatten_summary(_r)
+            return do_flatten_dict(_r)
 
         return _wrapper
 
     return _decorator
-
-
-def _do_flatten_summary(summary: t.Dict[str, t.Any]) -> t.Dict[str, t.Any]:
-    rt = {}
-
-    def _f_dict(_s: t.Dict[str, t.Any], _prefix: str = "") -> None:
-        for _k, _v in _s.items():
-            # TODO: waiting fix datastore bug
-            if _v is None:
-                continue
-            _k = f"{_prefix}{_k}"
-            if isinstance(_v, dict):
-                _f_dict(_v, _prefix=f"{_k}/")
-            elif isinstance(_v, (tuple, list)):
-                _f_list(_v, _prefix=f"{_k}/")
-            else:
-                rt[_k] = _v
-
-    def _f_list(
-        data: t.Union[t.Tuple[t.Any, ...], t.List[t.Any]], _prefix: str = ""
-    ) -> None:
-        index = 0
-        for _d in data:
-            if isinstance(_d, dict):
-                _f_dict(_d, _prefix=f"{_prefix}/")
-            elif isinstance(_d, (tuple, list)):
-                _f_list(_d, _prefix=f"{_prefix}/")
-            else:
-                rt[f"{_prefix}{index}"] = _d
-            index += 1
-
-    _f_dict(summary)
-    return rt
 
 
 def _calculate_roc_auc(
